Add support for namespace-filtered page generators.
Before, generators had to get data for /all/ namespaces, and these results
were filtered. This commit adds support for communicating wanted namespaces
with generators.

Instead of creating a generator by passing a list of namespaces, a /function/
that returns a list is passed. This function is only called once the generator
starts generating Page objects, which is /after/ the GeneratorFactory has
completed parsing all command line arguments. Then, the getNamespaces function
is called, and the correct namespaces are used.



git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/trunk@11383 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6